auxiliary-app/requirements.txt
psutil
pycryptodome
cryptography
